requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.15,<3.16
django-environ>=0.11,<1
django-filter>=24.1,<25
whitenoise>=6.7,<7
gunicorn>=23,<24
psycopg[binary]>=3.2,<4
pyyaml>=6.0,<7
